backend/requirements.txt
faster-whisper>=0.10.0
RealtimeSTT>=0.1.12
pyaudio>=0.2.13
numpy>=1.24.0
ollama>=0.1.0
python-dotenv>=1.0.0
psutil>=5.9.0
requests>=2.31.0
pyyaml>=6.0

# New dependencies for Phase 1
pyyaml>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
